mdm_comics_backend/requirements.txt
# MDM Comics Backend
# FastAPI Stack

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Image Processing (for grade estimator mock + cover search)
pillow>=10.4.0
numpy==1.26.3
imagehash>=4.3.1

# Storage (S3 compatible)
boto3==1.34.14

# HTTP Client
httpx==0.26.0

# Web Scraping
beautifulsoup4==4.12.3

# User Agent Parsing (for analytics)
user-agents==2.2.0

# Payments
stripe==7.0.0

# Rate Limiting (P1-3)
slowapi==0.1.9

# Redis (for webhook idempotency, caching)
redis>=5.0.0

# Auth
bcrypt==4.0.1

# UPS Shipping Integration v1.28.0
cryptography>=41.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Code formatting (dev only)
# black==24.1.0
# isort==5.13.2

# ML Dependencies - uncomment when model is ready
# torch==2.6.0
# torchvision==0.21.0
# opencv-python==4.9.0.80
# scikit-learn==1.4.0

# ===== OUTREACH SYSTEM v1.5.0 =====

# Job Queue
arq>=0.25.0

# AI Content Generation
openai>=1.0.0

# Template Rendering
jinja2>=3.1.0
